moved checkAlias to commonBotUtils for shared usage.
